travel-data-reader/requirements.txt
git+https://github.com/alexprengere/neobase.git
pytest
confluent-kafka
psycopg2-binary